evaluation/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
albumentations>=1.3.0
tqdm>=4.65.0
pillow>=10.0.0
scipy>=1.11.0
plotly>=5.15.0
jupyter>=1.0.0
pyyaml>=6.0
umap-learn>=0.5.3
